Add the ability to include expected results in a vt session log
This CL adds the ability to include a header block in a recorded vt session.

The header block lists interestin offsets in the data, what the terminal should look like at those offsets, and where the cursor should be.

The vtscope.py tool is modified to read the list of offsets from the header and make them available as shortcuts to the 'seek' command.  This makes it easier to jump directly to an interesting place in the data without resorting to sticky notes.

Additionally, a new test suite is added which can replay recorded sessions.  It stops at all defined offsets and ensures that the terminal contains the correct data and that the cursor has been left in the same position.

BUG=chromium-os:24772
TEST=test_harness.html, 53/53 tests passed.

Change-Id: I977c2c04fcd1d80bd83e4368922a32256d203af6
Reviewed-on: https://gerrit.chromium.org/gerrit/13848
Commit-Ready: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>