hterm: Create component-extension version of crosh+hterm.
This adds a new crosh_builtin asset which is a snapshot of hterm
with the ssh portions removed.  It follows the same pattern as
chrome vox, which packages a zip version of the extension
for use as a default component extension.

BUG=chromium-os:27663,chromium-os:27630

TEST=Test that ctrl-alt-t opens crosh even with no extensions
installed.
TEST=Install hterm, test that ctrl-alt-t opens hterm instead of
component extension.
TEST=Disable hterm, ensure that ctrl-alt-t opens component
extension again.

Note: Use document.location.href from the js console to determine
which extension is launched for ctrl-alt-t.

Change-Id: Id8a4a109f74aa0b1a5c47a870d96437ad5d74e77
Reviewed-on: https://gerrit.chromium.org/gerrit/18899
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Commit-Ready: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>