hterm: Reject invalid resizes
There appears to be a race condition during load of a second
hterm window which causes us to try to realize a negative terminal
size.  This causes a hang (I haven't diagnosed exactly why).

The patch works around the issue by rejecting any attempt to resize
to or realize a column or row count <= 0.

BUG=chromium-os:34123
TEST=Start secure shell, connect to a host, press ctr-shift-n and
  check that the new window opens correctly.  Close new window and
  repeat a few times to make sure.
TEST=test_harness.html, 65/65 tests passed.

Change-Id: Ic04fc6edf002e1558a11aa919ad97a842b2b620d
Reviewed-on: https://gerrit.chromium.org/gerrit/32182
Reviewed-by: Marius Schilder <mschilder@google.com>
Commit-Ready: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>