hterm: options: add a slider for control alpha channels of colors
In the future, we should get transparent backgrounds, so in anticipation
of that, allow people to control the alpha channels of colors via range
input elements.

For now, it really only helps with the cursor.

While it doesn't make too much sense to apply this to foreground colors,
it doesn't hurt and keeps us from special casing things.

BUG=None
TEST=loaded option page and played with sliders and watched alpha change

Change-Id: Icfd0f904361ace8d2ad6f1b11c19a0d7940dff20
Reviewed-on: https://gerrit.chromium.org/gerrit/41141
Reviewed-by: Robert Ginda <rginda@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>