Small changes to make it so hterm works better in IE.  There are still some issues around cut-copy-paste.
Also exposes terminal's pref_ object in a getter, and has the terminal pass
itself in the callback in the runCommandClass method.  This makes it so
integrating code doesn't have to directly access what looks like a private
variable.

Also fixes issue with double pasting the text with ctrl-v.  This was happening
because we are now listening for textInput events on the document body, rather
than just the x-screen.  The pasteInput text area would propagate a textInput
event up.  We are now catching this and stopping it.

Change-Id: I4c48f65cfc95cbd185083ea54f387229c76be983
Reviewed-on: https://chromium-review.googlesource.com/197372
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>