hterm: keyboard: make addBindings more user friendly
Passing symbol string names (e.g. 'Ctrl-Shift-P') only works with the
addBindings function which also requires you pass a list of keys.  By
moving the parsing logic into addBinding, we can bring symbolic name
support to it too.

So addBindings continues to work the way it did before, but now the
addBinding function can be called with either a symbolic string or
with a parsed keyPattern object.

We also extend the action interface so people can pass a function
directly down.  This makes it much easier to add custom bindings for
embedders in one-off fashions.

Change-Id: I9e0e1e55a316751461644943a5927e3c69fa2916
Reviewed-on: https://chromium-review.googlesource.com/389491
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>