hterm: fix alt key tracking
We track the status of the alt key to so we can undo some keyboard
shortcuts (like "alt-backspace-is-meta-backspace" for Chrome OS).
In order to track alt+tab (which normally changes windows), we use
the blur event.  This ends up not working because:
- the blur event does not bubble
- we set the listener on the body element
- the x-screen element is the one that has focus

Thus when we switch away, the blur event fires for x-screen only
and we never see it.  This manifests as the alt key being "stuck":
- enable alt-backspace-is-meta-backspace
- type some text in a bash shell prompt
- alt+tab to a different window
- alt+tab back to Secure Shell
- try pressing the delete key and see it act as alt+backspace

Instead, let's use the focusout event to track the focus state.
This event bubbles, so the body element is properly notified.

BUG=chromium:704468

Change-Id: I7990c6e8e7c7a70e052deaf41f72ebb05c46c692
Reviewed-on: https://chromium-review.googlesource.com/475471
Reviewed-by: Brandon Gilmore <varz@google.com>
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>