hterm: delay display of dialog until it finishes loading
Currently we display the dialog immediately after we've constructed it.
However, the dialog itself takes time to load resources (css & scripts),
and to render, and it hasn't even loaded its settings.  This causes some
UI jank as the user can see it load and elements shift around, and the
fields are populated.

Instead, let's hide the dialog until we've finished all the communication
and it's finished loading all of its settings.  Now the user should only
see the fully formed dialog.

Change-Id: If137cd926e36d9a60a19c734f184c81beecd543d
Reviewed-on: https://chromium-review.googlesource.com/509208
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>