hterm: switch to standard MouseEvent.button
The MouseEvent.which field has not been standardized, so switch to the
standard MouseEvent.button field.

Unfortunately the values don't map exactly -- they're off by 1.  For
.which, {1,2,3} => {left,middle,right}, but .button maps {0,1,2} =>
{left,middle,right}.  People who haven't changed the default (auto)
won't notice.  But people who have made a selection need to update
the option.  If they used this a lot, they should notice quickly.

BUG=chromium:725703

Change-Id: I0b7290f96cc6dd4024a403eafaa9baf5373c0a77
Reviewed-on: https://chromium-review.googlesource.com/513564
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>