hterm: make keybinding parsing more resilient to user inputs
The "in" keyword in JS is a messy beast.  Most people expect it to only
match entries in a dict they've added, but it also matches all members
of objects they've inherited (including Object).  That means a binding
like "__proto__-A" or "toString-B" will pass the parser, but throw weird
errors later on as we try to execute the keys.

Leverage the hasOwnProperty helper to avoid these sort of errors.

BUG=chromium:722718

Change-Id: Icd359b00ff6ad1777520811ed576c9b4eee5cef1
Reviewed-on: https://chromium-review.googlesource.com/516746
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>