hterm: make keybindings more flexible
There's no good reason to require exact capitalization with keys in
most cases.  We don't differentiate between "Ctrl" or "CTRL", so let
users roll with either by normalizing user inputs to uppercase.

One could make an argument about confusing "A", "a", "Shift-a", and
"Shift-A", but we already have that to some degree today with just
"A" and "Shift-A".  The benefits for other keys are outweighed imo
as requiring people to always write in uppercaps is not natural when
we require the modifiers to be mixed caps.  e.g. Ctrl-SPACE is OK,
but CTRL-Space is invalid?

We also add aliases for a lot of keys with longer names.  Many of
these keys don't have dominant forms, so accept e.g. Escape and Esc.

BUG=chromium:722718

Change-Id: Ib80a2b65f01fab2fbbe100c53deb61e9217b9358
Reviewed-on: https://chromium-review.googlesource.com/516747
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>