hterm: character maps: make override handling cumulative
The current override logic ends up clobbering the entire map.  So if
someone wants to tweak just one mapping for an existing character map,
they have to duplicate the entire thing.

Let's change the override logic so that it's cumulative: if the user
wants to change just "a"->"A", that now happens without the rest of
the character map being thrown away.

This splits out the reset logic into more reasonable behavior: it
actually resets the object state back to a pristine state.  This also
breaks the API, but we haven't been expecting anyone to use this
directly, so hopefully no one even notices.

This also makes the character overriding work in a few more scenarios.
Notably, when the map is null like the US map.

BUG=chromium:718670

Change-Id: I8771b3f78ddb7395ec0d7cdb8f194abb35b6682e
Reviewed-on: https://chromium-review.googlesource.com/539176
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>