hterm: fix mouse move reporting
The recent change 2edd36139eb27918c8f98e7430adca8126026630 (hterm:
switch to standard MouseEvent.button) broke reporting when moving
the mouse.  Before, we'd only report when a button was pushed, but
now we incorrectly reported even when no buttons are pushed.

Also take the opporunity to fix this code so that it better complies
with the xterm mouse protocol.  Namely, the middle button is higher
precedence than the right button, and we handle higher buttons.

BUG=chromium:733413

Change-Id: I70a8cffb73230e0bc4671242eeb954e279e5774e
Reviewed-on: https://chromium-review.googlesource.com/536355
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>