hterm: character maps: drop support for GR maps
GR map handling has broken since it was first added 5 years ago.  Fixing
computation of grmap in hterm.VT.CharacterMap is easy, but that leads to
another issue -- hterm.VT.parseUnknown_ does double translation on strings
akin to self.terminal.print(GR(GL(str))).  Fixing that would require a bit
more rework in a few places.  While feasible, let's just drop it instead.
No one has complained, and the usefulness of a GR map in a UTF8 world is
questionable at best.  When processing a stream of codeunits, it's easy to
mangle just one of them leading to broken output.

If anyone asks for it in the future, we can reconsider how to process the
content correctly (and add some tests!).

BUG=chromium:734318

Change-Id: I0006b5360cb0516065660d9f4c2e734bc3355168
Reviewed-on: https://chromium-review.googlesource.com/538831
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>