hterm: make virtual keyboards show up
Currently browsers don't show the virtual keyboard when the terminal has
focus.  This is because the way we capture keystrokes doesn't inform the
browser that we care -- we listen for keypress events and such.

Lets try marking the x-screen node as editable, and then turning off all
the editing related settings.  Since the terminal cursor is controlled by
us to indicate location, and we fully capture keystrokes, this should be
sufficient.

BUG=chromium:402972
TEST=loaded crosh on a cyan and put it into tablet mode, then switched tabs/focus
TEST=loaded test page on macOS and it still worked fine
TEST=loaded test page on Android/Chrome and the keyboard came up

Change-Id: I12e17d3b79c03e99a5d36fe7a199881952d799d9
Reviewed-on: https://chromium-review.googlesource.com/547003
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>