hterm: fill out escape sequence docs better
We document all the escape sequences that overlap with the C1 control
space in the C1 section, but then the dedicated escape section either
omits them or incorrectly duplicates (and marks them as ignored).  Fill
out the escape table so all characters are listed and we point people
to the C1 section for all the relevant sequences.

Change-Id: I0f46cc1a2d299db454b66a4f9da376430b4b91c7
Reviewed-on: https://chromium-review.googlesource.com/559894
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>