nassh: shrink & shift connection dialog
The current connection dialog uses flex boxes and css to always fill up
80% of the vertical space.  When you have a tall terminal, this can lead
to a lot of dead space between the connection list, settings, and buttons.

Lets rework the dialog so that it is only as tall as it needs to be to
display all the content without scrollbars.  We add max/min height limits
via css so we don't have to worry about pathological cases.

Finally, with this more compact dialog, we can shift it from always being
near the top to being near the bottom.  This frees up space at the top so
our FAQ/tips/notices we display are more frequently readable.  The current
layout ends up overlapping that content all the time.

Change-Id: Ia52010d41cec1d6d4e91915daa9a40ab58ae9fac
Reviewed-on: https://chromium-review.googlesource.com/581912
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>