hterm: avoid recalculating string display widths when possible
There's a number of locations where we calculate a string's display width,
pass the string to another function, and then that function also calculates
the display width.  Since wcwidth has to iterate over every codepoint every
time, this can be fairly wasteful.  So when we've already know the display
width ahead of time, pass the result around.  This covers the main loops in
the hterm display code.

Using Chrome's built in profiling, lib.wc.* funcs go from the top of the
profile (hottest funcs) to the 3rd hottest and lower.  That means rendering
of lots of text (ASCII & UTF-8) should get a decent boost.

Change-Id: I906c745b6e35d1990eebdc8436c0bce3a3ab8f4d
Reviewed-on: https://chromium-review.googlesource.com/634161
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>