hterm: do not warn about unknown sequences by default
Dumping this info was probably useful in the earlier hterm days, but now
that things are largely stable, this is a liability.  If you dump a file
with a lot of binary content (on purpose or by accident), generating a
log message per byte can easily overwhelm Chrome.  A simple example is
to dump /dev/zero.  Even just 10k bytes can start choking, and `cat` on
that can write MBs per second.

BUG=chromium:220406

Change-Id: I3295cd220fbedffbcc8b065b74ceaeb419cd59c3
Reviewed-on: https://chromium-review.googlesource.com/634744
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>