hterm: fix replacing wide chars with narrow chars
We handled backing up over wide chars when processing them in their
entirety, but our edge case checking missed deleting a single column
in the right half of the character (where offset is not zero in the
first loop iteration).  The count math would then go negative and
we'd loop forever as we only stop when count hits zero.

Add another edge case check for this, and add a sanity check for the
count so we at least don't loop forever if there's another bug.

We change an existing test's expectations to match the new behavior:
if we have a wide character, and the cursor is in the middle of it,
and we issue an erase-to-the-right command, the first column of the
wide character should be turned into a space.  Before, we would also
delete it.  Logically this makes more sense as the cursor location
is a fixed coordinate, so deleting a wide character should not cause
it to also move to the left by a column.

BUG=chromium:577691

Change-Id: Iafc3510489dcc88f4284d4cbea663b4b973fc1b7
Reviewed-on: https://chromium-review.googlesource.com/641372
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>