hterm: add support for a context menu
This adds a basic UI for people to Ctrl-Right click on the terminal
and get a bunch of common options.  hterm itself doesn't provide a
default menu currently, so the callers need to specify it entirely.

For now, we only support a flat menu.  We might want to explore
nested menus in the future if we want more stuff in here.

Bug: chromium:213983
Change-Id: I5a58d135ded1dc214e628d556213cb46fa5bc3a6
Reviewed-on: https://chromium-review.googlesource.com/1203630
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>