hterm: io: add a hideOverlay helper
Sometimes we want to show the overlay for an event with an unknown end
time.  Currently the only way to hide the overlay is to use the show
overlay function with a short timeout (1 msec) which makes the API a
bit hacky.  Split the existing hiding logic out and into a dedicated
function for people to call.

Change-Id: Ib79bd40d47a974bce2313aede8966b8437848e59
Reviewed-on: https://chromium-review.googlesource.com/671646
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>