hterm: io: buffer data when in background
The IO model makes it easy to add a new foreground process/command that
gets full control.  However, it requires all background processes to be
silent for the duration.  When dealing with asynchronous processes, this
requirement can be hard to satisfy without making every process check to
see if it's been placed in the background and if so, queue new data.  We
can however add this logic in the IO object itself so all users get it
for free.  Now all background data is queued up and once the IO object
is moved to the foreground, all that gets flushed at once.

Change-Id: Id96c526b40916240797a11fb1cad1dadfa0a3b88
Reviewed-on: https://chromium-review.googlesource.com/671647
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>