hterm: fix handling of inverse text with true colors
The text attribute code would update the foreground/background colors
only when they weren't already in true color form.  This was somewhat
intentional in that updates to the color (like bolding and faint) are
for non-true color sources, but it unintentionally missed flipping the
foreground & background colors when inverse is active.

We fix this by not writing directly to the foreground/background attrs
when setting true colors.  This lets us always treat them as outputs
by the sync code.  Instead, we continue using the "source" attrs like
we do for the palette based colors.  This lets us drop the SRC_RGB
constant as there are no more users.

BUG=chromium:736025

Change-Id: Id10ffcefcbac69d0d920de12cac910065d8902a5
Reviewed-on: https://chromium-review.googlesource.com/765590
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>