hterm: return early on mouse move events when possible
Add an explicit check for when mouse drag reporting is disabled, but
mouse click reporting is enabled.  Otherwise we do a bunch of early
event processing whenever the mouse moves across the screen even if
we don't actually report it.  We special case mouse movement as it
is a pretty common event and it makes debugging non-mouse moves a
bit of a pita.

Change-Id: I7f82a66348f1d1f93523a773b1b69173464c26c6
Reviewed-on: https://chromium-review.googlesource.com/764752
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>